pyproject.toml
[tool.poetry]
name = "pile"
version = "0.1.0"
description = ""
authors = ["Tomáš Mládek <t@mldk.cz>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
django-model-utils = "^4.0.0"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
